Build 3.3.6301.19490
* Updated to YC 119.3 Database and Images
* Fixed a Shopping List bug that would display the incorrect number of
R.A.Ms to build (runs) in the build list
* Fixed a Shopping List bug that would not correctly update amounts in
the buy list when updating runs in the Item List
* Fixed a Shopping List bug that deleted items from the buy list when
some build items were updated
* Items in Stock and Items in Production columns are visible on the
Manufacturing Tab again.
* Saving settings in the Industry Belt Upgrades tool will no longer
throw a save settings error.
* Double-clicking build items in the BP component list will now load the
runs to get the quantity shown in the grid (490 runs of missiles vs.
49,000 runs to get 49,000 missiles)
* Removed the teams box from the BP tab and put BP options in its place.
These options were in the 'Options' tab on facilities prior to this
change.